django/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/base.in
#
anyascii==0.1.7
    # via wagtail
argon2-cffi==20.1.0
    # via -r requirements/base.in
asgiref==3.3.1
    # via django
beautifulsoup4==4.8.2
    # via wagtail
certifi==2020.12.5
    # via
    #   elasticsearch
    #   requests
cffi==1.14.4
    # via argon2-cffi
chardet==4.0.0
    # via requests
defusedxml==0.6.0
    # via python3-openid
django==3.0.10
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-appconf
    #   django-controlled-vocabulary
    #   django-cors-headers
    #   django-filter
    #   django-formtools
    #   django-model-utils
    #   django-nine
    #   django-redis
    #   django-reversion
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-allauth==0.42.0
    # via -r requirements/base.in
django-appconf==1.0.4
    # via django-compressor
django-compressor==2.4
    # via
    #   -r requirements/base.in
    #   django-libsass
django-controlled-vocabulary==0.11
    # via django-ead
django-cors-headers==3.5.0
    # via -r requirements/base.in
django-crispy-forms==1.9.2
    # via -r requirements/base.in
django-ead==0.8.1
    # via -r requirements/base.in
django-elasticsearch-dsl==7.1.4
    # via
    #   -r requirements/base.in
    #   django-elasticsearch-dsl-drf
django-elasticsearch-dsl-drf==0.21
    # via -r requirements/base.in
django-environ==0.4.5
    # via -r requirements/base.in
django-filter==2.4.0
    # via wagtail
django-formtools==2.2
    # via -r requirements/base.in
django-kdl-wagtail==0.9.7
    # via -r requirements/base.in
django-libsass==0.8
    # via -r requirements/base.in
django-model-utils==4.0.0
    # via
    #   -r requirements/base.in
    #   django-ead
django-modelcluster==5.1
    # via wagtail
django-nine==0.2.4
    # via django-elasticsearch-dsl-drf
django-redis==4.12.1
    # via -r requirements/base.in
django-reversion==3.0.9
    # via django-ead
django-taggit==1.3.0
    # via wagtail
django-treebeard==4.4
    # via wagtail
djangorestframework==3.11.1
    # via
    #   -r requirements/base.in
    #   django-elasticsearch-dsl-drf
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
elasticsearch==7.10.1
    # via
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
elasticsearch-dsl==7.3.0
    # via
    #   -r requirements/base.in
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
et-xmlfile==1.0.1
    # via openpyxl
hiredis==1.1.0
    # via -r requirements/base.in
html5lib==1.1
    # via wagtail
idna==2.10
    # via requests
jdcal==1.4.1
    # via openpyxl
l18n==2020.6.1
    # via wagtail
libsass==0.20.1
    # via django-libsass
lxml==4.6.2
    # via django-ead
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==3.0.6
    # via tablib
pillow==7.2.0
    # via
    #   -r requirements/base.in
    #   wagtail
pycparser==2.20
    # via cffi
python-dateutil==2.8.1
    # via elasticsearch-dsl
python-slugify==4.0.1
    # via -r requirements/base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2020.1
    # via
    #   -r requirements/base.in
    #   django
    #   django-modelcluster
    #   l18n
rcssmin==1.0.6
    # via
    #   -r requirements/base.in
    #   django-compressor
redis==3.5.3
    # via
    #   -r requirements/base.in
    #   django-redis
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
    #   wagtail
requests-oauthlib==1.3.0
    # via django-allauth
rjsmin==1.1.0
    # via django-compressor
six==1.15.0
    # via
    #   argon2-cffi
    #   django-compressor
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
    #   html5lib
    #   l18n
    #   libsass
    #   python-dateutil
soupsieve==2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
tablib[xls,xlsx]==3.0.0
    # via wagtail
text-unidecode==1.3
    # via python-slugify
urllib3==1.26.3
    # via
    #   django-controlled-vocabulary
    #   elasticsearch
    #   requests
wagtail==2.12
    # via
    #   -r requirements/base.in
    #   django-kdl-wagtail
webencodings==0.5.1
    # via html5lib
whitenoise==5.2.0
    # via -r requirements/base.in
willow==1.4
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==1.3.7
    # via wagtail
xlwt==1.3.0
    # via tablib
